Revert "refactor(mcp): migrate include_data server to new architecture (#20511)"
This reverts commit 77008ee70793d7c600fd639977017ff1418916bf.
