--- front/lib/api/actions/servers/include_data/helpers.ts
@@ -1,66 +1,0 @@
-// eslint-disable-next-line dust/enforce-client-types-in-public-api
-import { INTERNAL_MIME_TYPES } from "@dust-tt/client";
-
-import type {
-  IncludeResultResourceType,
-  WarningResourceType,
-} from "@app/lib/actions/mcp_internal_actions/output_schemas";
-import { renderRelativeTimeFrameForToolOutput } from "@app/lib/actions/mcp_internal_actions/rendering";
-import {
-  getDataSourceNameFromView,
-  getDisplayNameForDocument,
-} from "@app/lib/data_sources";
-import type { CoreAPIDocument, TimeFrame } from "@app/types";
-import { stripNullBytes } from "@app/types";
-import type { DataSourceViewType } from "@app/types/data_source_view";
-
-export function makeIncludeWarningResource(
-  documents: CoreAPIDocument[],
-  topK: number,
-  timeFrame: TimeFrame | null
-): WarningResourceType | null {
-  const timeFrameAsString = renderRelativeTimeFrameForToolOutput(timeFrame);
-
-  // Check if the number of chunks reached the limit defined in params.topK.
-  const tooManyChunks =
-    documents &&
-    documents.reduce((sum, doc) => sum + doc.chunks.length, 0) >= topK;
-
-  // Determine the retrieval date limit from the last document's timestamp.
-  const retrievalTsLimit = documents?.[documents.length - 1]?.timestamp;
-  const date = retrievalTsLimit ? new Date(retrievalTsLimit) : null;
-  const retrievalDateLimitAsString = date
-    ? `${date.toLocaleString("default", { month: "short" })} ${date.getDate()}`
-    : null;
-
-  return tooManyChunks
-    ? {
-        mimeType: INTERNAL_MIME_TYPES.TOOL_OUTPUT.WARNING,
-        warningTitle: `Only includes documents since ${retrievalDateLimitAsString}.`,
-        warningData: { includeTimeLimit: retrievalDateLimitAsString ?? "" },
-        text: `Warning: could not include all documents ${timeFrameAsString}. Only includes documents since ${retrievalDateLimitAsString}.`,
-        uri: "",
-      }
-    : null;
-}
-
-export function makeIncludeResultResource(
-  doc: CoreAPIDocument,
-  dataSourceView: DataSourceViewType,
-  refs: string[]
-): IncludeResultResourceType {
-  return {
-    mimeType: INTERNAL_MIME_TYPES.TOOL_OUTPUT.DATA_SOURCE_INCLUDE_RESULT,
-    uri: doc.source_url ?? "",
-    text: getDisplayNameForDocument(doc),
-
-    id: doc.document_id,
-    source: {
-      provider: dataSourceView.dataSource.connectorProvider ?? undefined,
-      name: getDataSourceNameFromView(dataSourceView),
-    },
-    tags: doc.tags,
-    ref: refs.shift() as string,
-    chunks: doc.chunks.map((chunk) => stripNullBytes(chunk.text)),
-  };
-}
